+++ tests/models/test_embeddables.py
@@ -3,6 +3,7 @@ from typing import ClassVar
 import pytest
 
 import edgy
+from edgy.core.db.fields.base import BaseField
 from edgy.testclient import DatabaseTestClient as Database
 from tests.settings import DATABASE_URL
 
@@ -34,7 +35,9 @@ class MyModel1(edgy.Model):
 
 class MyModel2(MyModel1):
     model3: ClassVar[MyModel1] = MyModel1
+    # because of exluding and inheritance we cannot use model3 anymore
     model3_model1_last_name = edgy.ExcludeField()
+
 
 
 @pytest.fixture()
@@ -70,6 +73,31 @@ def test_fields():
     assert isinstance(MyModel2.meta.fields_mapping["model3_model1_last_name"], edgy.ExcludeField)
 
 
+@pytest.mark.parametrize(
+    "model",
+    [
+        MyModel1, MyModel2
+    ],
+)
+def test_field_types(model):
+    for field in model.meta.fields_mapping.values():
+        assert not isinstance(field, edgy.Manager)
+    for field in model.meta.fields_mapping.values():
+        assert isinstance(field, BaseField)
+    for manager in model.meta.managers.values():
+        assert isinstance(manager, edgy.Manager)
+
+
+@pytest.mark.parametrize(
+    "model",
+    [
+        MyModel1, MyModel2
+    ],
+)
+def test_field_names(model):
+    for field_name, field in model.meta.fields_mapping.items():
+        assert field_name == field.name
+
 async def test_fields_db(rollback_connections):
     model1_def = {
         "model1_first_name": "edgy",
@@ -96,24 +124,19 @@ async def test_fields_db(rollback_connections):
             "last_name": "edgytoo",
         },
         "model3": {
-            "model1_first_name": "edgy",
-            "model2_last_name": "edgy",
+            "model1": {
+                "first_name": "edgy",
+                "last_name": "edgytoo",
+            },
         }
     }
 
 
-    # TODO: fix failure for embedded Composite
-    # stub for preventing mypy to complain
-    assert model2_def == model2_def
+    obj2 = await MyModel2.query.create(**model2_def)
 
-    #obj2 = await MyModel2.query.create(**model2_def)
-
-    #assert obj.model_dump() == {
-    #    "model1": {
-    #        "first_name": "edgy",
-    #        "last_name": "ed2gytoo",
-    #    },
-    #    "model3": {
-    #        "first_name": "edgy",
-    #    }
-    #}
+    # dumping model3 is broken because some fields are missing for Constructors
+    assert obj2.model_dump(include={"model1": True}) == {"model1": model2_def["model1"]}
+    for key, val in (("model1_first_name", "edgy"),("model1_last_name", "edgytoo"),("model3_model1_first_name", "edgy")):
+        assert getattr(obj2, key) == val
+    with pytest.raises(AttributeError):
+        obj2.model3_model2_first_name  # noqa